+++ src/app/notes-manager/layout.tsx
